Add advisor template data request
